@@ -20,73 +24,67 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "~14.1.3",
-    "@angular/cdk": "~14.1.3",
-    "@angular/common": "~14.1.3",
-    "@angular/compiler": "~14.1.3",
-    "@angular/core": "~14.1.3",
-    "@angular/flex-layout": "^14.0.0-beta.40",
-    "@angular/forms": "~14.1.3",
-    "@angular/localize": "~14.1.3",
-    "@angular/material": "~14.1.3",
-    "@angular/platform-browser": "~14.1.3",
-    "@angular/platform-browser-dynamic": "~14.1.3",
-    "@angular/router": "~14.1.3",
-    "@angular/service-worker": "~14.1.3",
-    "@ngx-translate/core": "^14.0.0",
+    "@angular/animations": "18.2.2",
+    "@angular/cdk": "18.2.2",
+    "@angular/common": "18.2.2",
+    "@angular/compiler": "18.2.2",
+    "@angular/core": "18.2.2",
+    "@angular/forms": "18.2.2",
+    "@angular/localize": "18.2.2",
+    "@angular/platform-browser": "18.2.2",
+    "@angular/platform-browser-dynamic": "18.2.2",
+    "@angular/router": "18.2.2",
+    "@angular/service-worker": "18.2.2",
     "lodash": "^4.17.20",
-    "material-design-icons-iconfont": "^6.1.0",
-    "rxjs": "^7.5.0",
+    "rxjs": "7.8.1",
     "tailwindcss": "^3.4.4",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.1.3",
-    "@angular-eslint/builder": "~14.0.3",
-    "@angular-eslint/eslint-plugin": "~14.0.3",
-    "@angular-eslint/eslint-plugin-template": "~14.0.3",
-    "@angular-eslint/schematics": "~14.0.3",
-    "@angular-eslint/template-parser": "~14.0.3",
-    "@angular/cli": "~14.1.3",
-    "@angular/compiler-cli": "~14.1.3",
-    "@angular/language-service": "~14.1.3",
-    "@biesbjerg/ngx-translate-extract": "^7.0.3",
-    "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
+    "@angular-devkit/build-angular": "18.2.2",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/cli": "18.2.2",
+    "@angular/compiler-cli": "18.2.2",
+    "@angular/language-service": "18.2.2",
     "@compodoc/compodoc": "^1.1.11",
     "@ngneat/until-destroy": "^9.0.0",
-    "@ngx-rocket/scripts": "^5.2.2",
     "@types/jasmine": "^4.0.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/lodash": "^4.14.184",
-    "@types/node": "^14.18.26",
-    "@typescript-eslint/eslint-plugin": "~5.34.0",
-    "@typescript-eslint/parser": "~5.34.0",
+    "@types/node": "18.19.47",
+    "@typescript-eslint/eslint-plugin": "8.3.0",
+    "@typescript-eslint/parser": "8.3.0",
+    "daisyui": "^4.12.10",
     "eslint": "^8.3.0",
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
-    "hads": "^3.0.0",
     "https-proxy-agent": "^5.0.0",
     "husky": "^8.0.1",
-    "jasmine-core": "~4.2.0",
-    "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
+    "jasmine-core": "4.2.0",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "6.4.0",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage": "2.2.0",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.0.0",
     "karma-junit-reporter": "^2.0.1",
+    "material-icons": "^1.13.12",
     "postcss": "^8.4.5",
     "prettier": "^2.2.1",
     "pretty-quick": "^3.1.0",
     "puppeteer": "^16.2.0",
-    "stylelint": "~14.11.0",
+    "stylelint": "14.11.0",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended-scss": "~7.0.0",
-    "stylelint-config-standard": "~28.0.0",
+    "stylelint-config-recommended-scss": "7.0.0",
+    "stylelint-config-standard": "28.0.0",
     "ts-node": "^10.1.0",
-    "typescript": "~4.7.4"
+    "typescript": "5.4.5"
   },
   "prettier": {
     "singleQuote": true,
