@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { AboutComponent } from './about.component';
 
-import { HomeComponent } from './home.component';
-
-const routes: Routes = [{ path: '', component: HomeComponent, title: 'klesiewicz.dev | home' }];
+const routes: Routes = [{ path: 'about', component: AboutComponent, title: 'klesiewicz.dev | about' }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
   providers: [],
 })
-export class HomeRoutingModule {}
+export class AboutRoutingComonent {}
