@@ -4,5 +4,18 @@ module.exports = {
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require('daisyui')],
+  daisyui: {
+    themes: [
+      {
+        mytheme: {
+          primary: '#050024',
+          secondary: '#7e6300',
+          accent: '#600106',
+          neutral: '#ece4dc',
+          'base-100': '#ece4dc',
+        },
+      },
+    ],
+  },
 };
